Revert "Simplifying how a GLMManager adds a substation meter, since the previous"
The GLMManager originally disconnected any lines out of the substation and reconnected
them to the new feeder. The 9500 node model brought up a bug in GridLAB-D. See
https://github.com/gridlab-d/gridlab-d/issues/1197#issue-506615238.

This GridLAB-D issue is now fixed. If the line is not connected to the meter, nothing
flows through the meter and results in 0 measurements (as described here:
https://github.com/gridlab-d/gridlab-d/issues/1198#issue-506736957). So, this commit
needs undone. I'll manually add back in the new model and associated tests, as they're
worth keeping.

This reverts commit 71478d983970ea6fb813b44a425c35fd61ae9d79.